kambari-agent/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
streamlit==1.30.0
sqlalchemy==2.0.25
python-dateutil==2.8.2
pandas==2.1.4
click==8.1.7

# Google Gemini AI
google-generativeai==0.3.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Database
alembic==1.12.1

# Web Server
uvicorn==0.24.0
fastapi==0.104.1
python-multipart==0.0.6

# Utilities
python-slugify==8.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0